All required files for the exercise
