+++ pyproject.toml
@@ -1,0 +1,16 @@
+[build-system]
+requires = ["setuptools>=41", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "luxik_diffusion2d"
+version = "0.0.2"
+readme = "README.md"
+description = "Exercise for Simulation"
+requires-python = ">=3.8"
+dependencies = [
+    "numpy", "matplotlib"
+]
+
+[project.urls]
+Repository = "https://github.com/Simulation-Software-Engineering/diffusion2d"
